test(e2e_app): replace flaky Hacker News webview test with Patrol docs
The Hacker News login flow was the flaky webview dependency: after tapping
'login', the native automator frequently found no EditText to type into.
Replace it with a test against the Patrol docs (patrol.leancode.co):
open the sidebar, navigate to Getting started, then type into the search
input. Icon-only controls are matched by their aria-label (content
description on Android, accessibility label on iOS).
