+++ dev/e2e_app/lib/main.dart
@@ -269,12 +269,12 @@ class _ExampleHomePageState extends State<ExampleHomePage> {
             onPressed: () => Navigator.of(context).push(
               MaterialPageRoute<void>(
                 builder: (_) => const WebViewScreen(
-                  title: 'WebView (Hacker News)',
-                  url: 'https://news.ycombinator.com',
+                  title: 'WebView (Patrol docs)',
+                  url: 'https://patrol.leancode.co',
                 ),
               ),
             ),
-            child: const Text('Open webview (Hacker News)'),
+            child: const Text('Open webview (Patrol docs)'),
           ),
           TextButton(
             onPressed: () => Navigator.of(context).push(
